requirements.txt
flask>=3.0.0,<4
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0
pymupdf>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
